requirements.txt
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
gunicorn==19.9.0
packaging==17.1
olefile==0.45.1
psycopg2==2.7.5
Pillow==5.1.0
pkginfo==1.4.2
six==1.11.0
pyuca==1.2
django-storages==1.7.1
django-bmemcached==0.2.3
django-axes==4.4.0
django-translations==1.1.0
django-appconf==1.0.2
django-ipware==2.1.0
boto3==1.9.14
jmespath==0.9.3
python-dateutil==2.7.3
s3transfer==0.1.13

python-decouple==3.1


djangorestframework==3.8.2